requirements.txt
requests
beautifulsoup4
lxml
tldextract
python-dateutil
pandas
sqlalchemy
aiohttp
aiodns
async-timeout
validators
Pillow
selenium 
webdriver-manager
python-dotenv
scikit-learn 
joblib
pytest
numpy>=2.0
